# tests/app/__snapshots__/test_app/test_openapi_spec.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PluginDetails": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/ExtractorResponse"
          },
          {
            "$ref": "#/components/schemas/LoaderResponse"
          },
          {
            "$ref": "#/components/schemas/UtilityResponse"
          },
          {
            "$ref": "#/components/schemas/OrchestratorResponse"
          },
          {
            "$ref": "#/components/schemas/TransformResponse"
          },
          {
            "$ref": "#/components/schemas/TransformerResponse"
          },
          {
            "$ref": "#/components/schemas/MapperResponse"
          },
          {
            "$ref": "#/components/schemas/FileResponse"
          }
        ]
      },
      "PluginIndex": {
        "additionalProperties": {
          "$ref": "#/components/schemas/PluginTypeIndex"
        },
        "propertyNames": {
          "$ref": "#/components/schemas/PluginTypeEnum"
        },
        "type": "object"
      },
      "PluginListElement": {
        "additionalProperties": false,
        "description": "Plugin list element model.",
        "properties": {
          "plugin": {
            "description": "The plugin name",
            "examples": [
              "tap-github"
            ],
            "title": "Plugin",
            "type": "string"
          },
          "plugin_type": {
            "$ref": "#/components/schemas/PluginTypeEnum",
            "description": "The plugin type",
            "examples": [
              "extractors"
            ]
          },
          "ref": {
            "description": "API URL for the variant",
            "examples": [
              "/meltano/api/v1/plugins/extractors/tap-github--singer-io"
            ],
            "title": "Ref",
            "type": "string"
          },
          "variant": {
            "description": "The plugin variant",
            "examples": [
              "meltanolabs"
            ],
            "title": "Variant",
            "type": "string"
          }
        },
        "required": [
          "ref",
          "plugin",
          "variant",
          "plugin_type"
        ],
        "title": "PluginListElement",
        "type": "object"
      },
      "PluginRef": {
        "description": "Plugin entry.",
        "properties": {
          "default_variant": {
            "description": "The default variant of the plugin",
            "examples": [
              "singer-io"
            ],
            "title": "Default Variant",
            "type": "string"
          },
          "logo_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "URL to the plugin's logo",
            "title": "Logo Url"
          },
          "variants": {
            "additionalProperties": {
              "$ref": "#/components/schemas/VariantReference"
            },
            "description": "The variants of the plugin",
            "title": "Variants",
            "type": "object"
          }
        },
        "required": [
          "default_variant"
        ],
        "title": "PluginRef",
        "type": "object"
      },
...
      "PluginTypeEnum": {
        "description": "Plugin types.",
        "enum": [
          "extractors",
          "loaders",
          "transformers",
          "utilities",
          "transforms",
          "orchestrators",
          "mappers",
          "files"
        ],
        "title": "PluginTypeEnum",
        "type": "string"
      },
      "PluginTypeIndex": {
        "additionalProperties": {
          "$ref": "#/components/schemas/PluginRef"
        },
        "type": "object"
      },
      "PluginTypeOrAnyEnum": {
        "description": "One or any plugin type enum.",
        "enum": [
          "extractors",
          "loaders",
          "transformers",
          "utilities",
          "transforms",
          "orchestrators",
          "mappers",
          "files",
          "any"
        ],
        "title": "PluginTypeOrAnyEnum",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VariantReference": {
        "additionalProperties": false,
        "description": "Variant reference model.",
        "properties": {
          "ref": {
            "description": "API URL for the variant",
            "examples": [
              "/meltano/api/v1/plugins/extractors/tap-github--singer-io"
            ],
            "title": "Ref",
            "type": "string"
          }
        },
        "required": [
          "ref"
        ],
        "title": "VariantReference",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "The Meltano Hub API provides access to Meltano's plugin registry. It allows you to search for plugins, view their details, and download the necessary files to install them.\n\n- The API is versioned, with the current version being v1.\n- The API is read-only, and no authentication is required.\n",
    "title": "Meltano Hub API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/meltano/api/v1/plugins/index": {
      "get": {
        "description": "Retrieve global index of plugins.",
        "operationId": "get_plugin_index",
        "parameters": [
          {
            "description": "The `If-None-Match` HTTP request header makes the request conditional.\nFor `GET` and `HEAD` methods, the server will return the requested resource, with a `200` status, only if it doesn't have an `ETag` matching the given ones.\nFor other methods, the request will be processed only if the eventually existing resource's `ETag` doesn't match any of the values listed.\n\nhttps://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match\n",
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "description": "The `If-None-Match` HTTP request header makes the request conditional.\nFor `GET` and `HEAD` methods, the server will return the requested resource, with a `200` status, only if it doesn't have an `ETag` matching the given ones.\nFor other methods, the request will be processed only if the eventually existing resource's `ETag` doesn't match any of the values listed.\n\nhttps://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match\n",
              "pattern": "^\"etag-[0-9a-f]{16}\"$",
              "title": "If-None-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginIndex"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get plugin index",
        "tags": [
          "plugins"
        ]
      }
    },
    "/meltano/api/v1/plugins/made-with-sdk": {
      "get": {
        "description": "Retrieve plugins made with the Singer SDK.",
        "operationId": "get_sdk_plugins",
        "parameters": [
          {
            "description": "The number of plugins to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "description": "The number of plugins to return",
              "examples": [
                10,
                25,
                100
              ],
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "The plugin type",
            "in": "query",
            "name": "plugin_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PluginTypeOrAnyEnum",
              "default": "any",
              "description": "The plugin type",
              "examples": [
                "any",
                "extractors",
                "loaders"
              ]
            }
          },
          {
            "description": "The `If-None-Match` HTTP request header makes the request conditional.\nFor `GET` and `HEAD` methods, the server will return the requested resource, with a `200` status, only if it doesn't have an `ETag` matching the given ones.\nFor other methods, the request will be processed only if the eventually existing resource's `ETag` doesn't match any of the values listed.\n\nhttps://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match\n",
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "description": "The `If-None-Match` HTTP request header makes the request conditional.\nFor `GET` and `HEAD` methods, the server will return the requested resource, with a `200` status, only if it doesn't have an `ETag` matching the given ones.\nFor other methods, the request will be processed only if the eventually existing resource's `ETag` doesn't match any of the values listed.\n\nhttps://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match\n",
              "pattern": "^\"etag-[0-9a-f]{16}\"$",
              "title": "If-None-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PluginListElement"
                  },
                  "title": "Response Get Sdk Plugins",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get SDK plugins",
        "tags": [
          "plugins"
        ]
      }
    },
    "/meltano/api/v1/plugins/search": {
      "get": {
        "operationId": "get_plugin",
        "parameters": [
          {
            "description": "The plugin name",
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "description": "The plugin name",
              "examples": [
                "tap-github"
              ],
              "title": "Name",
              "type": "string"
            }
          },
          {
            "description": "The plugin type",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PluginTypeEnum",
              "description": "The plugin type",
              "examples": [
                "extractors"
              ]
            }
          },
          {
            "description": "The optional variant name",
            "in": "query",
            "name": "variant",
            "required": false,
            "schema": {
              "description": "The optional variant name",
              "examples": [
                "meltanolabs"
              ],
              "title": "Variant",
              "type": "string"
            }
          },
          {
            "description": "The `If-None-Match` HTTP request header makes the request conditional.\nFor `GET` and `HEAD` methods, the server will return the requested resource, with a `200` status, only if it doesn't have an `ETag` matching the given ones.\nFor other methods, the request will be processed only if the eventually existing resource's `ETag` doesn't match any of the values listed.\n\nhttps://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match\n",
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "description": "The `If-None-Match` HTTP request header makes the request conditional.\nFor `GET` and `HEAD` methods, the server will return the requested resource, with a `200` status, only if it doesn't have an `ETag` matching the given ones.\nFor other methods, the request will be processed only if the eventually existing resource's `ETag` doesn't match any of the values listed.\n\nhttps://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match\n",
              "pattern": "^\"etag-[0-9a-f]{16}\"$",
              "title": "If-None-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginDetails"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Not a valid plugin type"
          },
          "404": {
            "description": "Plugin not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Find a plugin",
        "tags": [
          "plugins"
        ]
      }
    },
    "/meltano/api/v1/plugins/stats": {
      "get": {
        "description": "Retrieve Hub plugin statistics.",
        "operationId": "get_plugin_stats",
        "parameters": [
          {
            "description": "The `If-None-Match` HTTP request header makes the request conditional.\nFor `GET` and `HEAD` methods, the server will return the requested resource, with a `200` status, only if it doesn't have an `ETag` matching the given ones.\nFor other methods, the request will be processed only if the eventually existing resource's `ETag` doesn't match any of the values listed.\n\nhttps://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match\n",
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "description": "The `If-None-Match` HTTP request header makes the request conditional.\nFor `GET` and `HEAD` methods, the server will return the requested resource, with a `200` status, only if it doesn't have an `ETag` matching the given ones.\nFor other methods, the request will be processed only if the eventually existing resource's `ETag` doesn't match any of the values listed.\n\nhttps://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match\n",
              "pattern": "^\"etag-[0-9a-f]{16}\"$",
              "title": "If-None-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "propertyNames": {
                    "$ref": "#/components/schemas/PluginTypeEnum"
                  },
                  "title": "Response Get Plugin Stats",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hub statistics",
        "tags": [
          "plugins"
        ]
      }
    },
    "/meltano/api/v1/plugins/{plugin_type}/index": {
      "get": {
        "description": "Retrieve index of plugins of a given type.",
        "operationId": "get_plugin_type_index",
        "parameters": [
          {
            "description": "The plugin type",
            "in": "path",
            "name": "plugin_type",
            "required": true,
            "schema": {
              "description": "The plugin type",
              "examples": [
                "extractors"
              ],
              "title": "Plugin Type",
              "type": "string"
            }
          },
          {
            "description": "The `If-None-Match` HTTP request header makes the request conditional.\nFor `GET` and `HEAD` methods, the server will return the requested resource, with a `200` status, only if it doesn't have an `ETag` matching the given ones.\nFor other methods, the request will be processed only if the eventually existing resource's `ETag` doesn't match any of the values listed.\n\nhttps://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match\n",
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "description": "The `If-None-Match` HTTP request header makes the request conditional.\nFor `GET` and `HEAD` methods, the server will return the requested resource, with a `200` status, only if it doesn't have an `ETag` matching the given ones.\nFor other methods, the request will be processed only if the eventually existing resource's `ETag` doesn't match any of the values listed.\n\nhttps://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match\n",
              "pattern": "^\"etag-[0-9a-f]{16}\"$",
              "title": "If-None-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginTypeIndex"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Not a valid plugin type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get plugin type index",
        "tags": [
          "plugins"
        ]
      }
    },
    "/meltano/api/v1/plugins/{plugin_type}/{plugin_name}--{plugin_variant}": {
      "get": {
        "description": "Retrieve details of a specific plugin variant.",
        "operationId": "get_plugin_variant",
        "parameters": [
          {
            "description": "The plugin type",
            "in": "path",
            "name": "plugin_type",
            "required": true,
            "schema": {
              "description": "The plugin type",
              "examples": [
                "extractors"
              ],
              "title": "Plugin Type",
              "type": "string"
            }
          },
          {
            "description": "The plugin name",
            "in": "path",
            "name": "plugin_name",
            "required": true,
            "schema": {
              "description": "The plugin name",
              "examples": [
                "tap-github"
              ],
              "title": "Plugin Name",
              "type": "string"
            }
          },
          {
            "description": "The plugin variant",
            "in": "path",
            "name": "plugin_variant",
            "required": true,
            "schema": {
              "description": "The plugin variant",
              "examples": [
                "meltanolabs"
              ],
              "title": "Plugin Variant",
              "type": "string"
            }
          },
          {
            "description": "The `If-None-Match` HTTP request header makes the request conditional.\nFor `GET` and `HEAD` methods, the server will return the requested resource, with a `200` status, only if it doesn't have an `ETag` matching the given ones.\nFor other methods, the request will be processed only if the eventually existing resource's `ETag` doesn't match any of the values listed.\n\nhttps://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match\n",
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "description": "The `If-None-Match` HTTP request header makes the request conditional.\nFor `GET` and `HEAD` methods, the server will return the requested resource, with a `200` status, only if it doesn't have an `ETag` matching the given ones.\nFor other methods, the request will be processed only if the eventually existing resource's `ETag` doesn't match any of the values listed.\n\nhttps://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match\n",
              "pattern": "^\"etag-[0-9a-f]{16}\"$",
              "title": "If-None-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginDetails"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Not a valid plugin type"
          },
          "404": {
            "description": "Plugin variant not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get plugin variant",
        "tags": [
          "plugins"
        ]
      }
    },
    "/meltano/api/v1/plugins/{plugin_type}/{plugin_name}/default": {
      "get": {
        "description": "Retrieve details of the default plugin variant.",
        "operationId": "get_default_plugin",
        "parameters": [
          {
            "description": "The plugin type",
            "in": "path",
            "name": "plugin_type",
            "required": true,
            "schema": {
              "description": "The plugin type",
              "examples": [
                "extractors"
              ],
              "title": "Plugin Type",
              "type": "string"
            }
          },
          {
            "description": "The plugin name",
            "in": "path",
            "name": "plugin_name",
            "required": true,
            "schema": {
              "description": "The plugin name",
              "examples": [
                "tap-github"
              ],
              "title": "Plugin Name",
              "type": "string"
            }
          },
          {
            "description": "The `If-None-Match` HTTP request header makes the request conditional.\nFor `GET` and `HEAD` methods, the server will return the requested resource, with a `200` status, only if it doesn't have an `ETag` matching the given ones.\nFor other methods, the request will be processed only if the eventually existing resource's `ETag` doesn't match any of the values listed.\n\nhttps://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match\n",
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "description": "The `If-None-Match` HTTP request header makes the request conditional.\nFor `GET` and `HEAD` methods, the server will return the requested resource, with a `200` status, only if it doesn't have an `ETag` matching the given ones.\nFor other methods, the request will be processed only if the eventually existing resource's `ETag` doesn't match any of the values listed.\n\nhttps://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match\n",
              "pattern": "^\"etag-[0-9a-f]{16}\"$",
              "title": "If-None-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "307": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Not a valid plugin type"
          },
          "404": {
            "description": "Plugin not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get the default plugin variant",
        "tags": [
          "plugins"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "http://localhost:8000"
    }
  ]
}